"check_style_score -> 1 of 1 passed. style score: 0/1 "

+++ assignments/m4/Assignment-3/Solution.java
@@ -1,7 +1,7 @@
 import java.util.Scanner;
 public final class Solution {
     static long binaryToDecimal(final String s) {
-        long i = (long) Long.parseLong(s);
+        long i = Long.parseLong(s);
         int b = 0;
         final int ten = 10;
         long sum = 0;
